Use string datatype for field Op in Error
- Remove the type xgo.Op.
- errors.WithOp({xgo.Op => string}).
- Change return type of *errors.Error.Ops() - {[]xgo.Op => []string}.
- Change data type of field Ops in errors.InternalDetails -
  {[]xgo.Op => []string}.

xgo.Op was not providing much value since it could not be directly
passed in as a constructor option. Dropping the type simplifies the API
and avoids having to import another package in some cases. It was
already possible to skip specifying the type in op declaration:

    const op = "JSONDecoder.Decode"
    //...
    return errors.E(errors.WithOp(op), errors.WithErr(err))

BREAKING CHANGE: The type xgo.Op has been removed. The following types
and functions now use the string datatype instead of xgo.Op:
- errors.WithOp(...)
- *errors.Error.Ops()
- errors.InternalDetails.Ops